[video] Separate VideoPlayer and VideoView into files
