+++ packages/expo-video/src/index.ts
@@ -1,6 +1,7 @@
 import Video from './NativeVideoModule';
 
-export { VideoView, useVideoPlayer, isPictureInPictureSupported } from './VideoView';
+export { VideoView, isPictureInPictureSupported } from './VideoView';
+export { useVideoPlayer } from './VideoPlayer';
 export { Video };
 export {
   VideoSource,
